tweak: rename and add role to user
