+++ src/restful/controllers/authController.js
@@ -198,6 +198,7 @@ class AuthController {
       const nonSensitiveData = {
         first_name: userData.first_name,
         last_name: userData.last_name,
+        role: userData.role,
         imageUrl: userData.imageUrl, // Assuming this field exists in the user document
         bio: userData.bio, // Assuming this field exists in the user document
       };
